remove old tuples builder #130
